fix(agents): filter undeliverable MCP servers at spawn, not only on save
Move MCP_SUPPORT/mcpAttachable from the editor's shared.ts into
data/providers.ts alongside isDockerSupported — it's provider capability
data, and the store layer needs it too. spawnFromDraft now drops servers
the spawning provider can't deliver (e.g. a pre-existing HTTP assignment
on a codex agent), so the session snapshot contains exactly what the
provider can run instead of ids the backend silently ignores.
